feat: split into mvc file directory
- Controller: RacingGame
- Domain: Car, Cars...
- View: InputView, OutputView
