+++ src/main/java/racingcar/domain/RuleStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface RuleStrategy {
     int getMovement();
