+++ src/main/java/racingcar/domain/WinStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface WinStrategy {
     Winners decideWinners(RoundRecords roundRecords);
