+++ src/main/java/racingcar/domain/WinStrategyImpl.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.LinkedHashMap;
 import java.util.List;
