+++ src/test/java/racingcar/GameManagerTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.*;
 
 import java.util.Arrays;
 import java.util.List;
