+++ src/test/java/racingcar/GameRoundsTest.java
@@ -6,6 +6,10 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
+import racingcar.domain.GameRounds;
+import racingcar.domain.RoundRecord;
 
 import java.util.Arrays;
 import java.util.List;
