+++ src/test/java/racingcar/RoundRecordsTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RoundRecord;
+import racingcar.domain.RoundRecords;
 
 import java.util.LinkedHashMap;
