+++ src/test/java/racingcar/WinStrategyImplTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.*;
 
 import java.util.LinkedHashMap;
